requirements.txt
# Core Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
tinydb>=4.8.0
aiohttp>=3.9.0
pydantic>=2.0.0
httpx>=0.25.0

# Frontend Dependencies  
streamlit>=1.28.0
plotly>=5.17.0

# Data Processing
pandas>=2.1.0
numpy>=1.25.0
yfinance>=0.2.18

# PDF Generation
reportlab>=4.0.0
matplotlib>=3.7.0

# Web Scraping
feedparser>=6.0.10
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.15.0
